unicorn config changed

+++ config/deploy.rb
@@ -28,18 +28,22 @@ set :unicorn_binary, "#{current_path}/bin/unicorn"
 set :unicorn_config, "#{current_path}/config/unicorn.rb"
 set :unicorn_pid, "#{current_path}/tmp/pids/unicorn.pid"
 
+def remote_file_exists?(full_path)
+  'true' ==  capture("if [ -e #{full_path} ]; then echo 'true'; fi").strip
+end
+
 namespace :deploy do
   task :start, :roles => :app, :except => { :no_release => true } do 
     run "cd #{current_path} && #{unicorn_binary} -c #{unicorn_config} -E #{rails_env} -D"
   end
   task :stop, :roles => :app, :except => { :no_release => true } do 
-    run "kill `cat #{unicorn_pid}`"
+    run "kill `cat #{unicorn_pid}`" if remote_file_exists?(unicorn_pid)
   end
   task :graceful_stop, :roles => :app, :except => { :no_release => true } do
-    run "kill -s QUIT `cat #{unicorn_pid}`"
+    run "kill -s QUIT `cat #{unicorn_pid}`" if remote_file_exists?(unicorn_pid)
   end
   task :reload, :roles => :app, :except => { :no_release => true } do
-    run "kill -s USR2 `cat #{unicorn_pid}`"
+    run "kill -s USR2 `cat #{unicorn_pid}`" if remote_file_exists?(unicorn_pid)
   end
   task :restart, :roles => :app, :except => { :no_release => true } do
     stop
